lib/requirements.txt
python-docx>=1.2.0
marko[codehilite]>=2.2.2
Pillow>=12.1.1
PyYAML>=6.0.3
rich>=13.0.0
